+++ inc/HM11/hm11.c
@@ -41,8 +41,8 @@ void Hm11_Packet(void){
     _tx_buffer_[14] = Mpu6050_.Temp_RAW >> 8;
     _tx_buffer_[15] = Mpu6050_.Temp_RAW;
     _tx_buffer_[16] = Mpu6050_.Sensitivity;
-    _tx_buffer_[17] = Mpu6050_.Num_packet >> 8;
-    _tx_buffer_[18] = Mpu6050_.Num_packet;
+    _tx_buffer_[17] = Hm11_.Num_packet >> 8;
+    _tx_buffer_[18] = Hm11_.Num_packet;
 
     _tx_buffer_[16] = _tx_buffer_[16] | (Hm11_.frequency<<6) ;
 
@@ -72,6 +72,7 @@ void Hm11_Config_By_Cmd(void){
                 break;
             case HM11_TX_OFF:
                 Hm11_.onoff = HM11_TX_OFF;
+                Hm11_.
                 break;
             case HM11_SET_TX_FREQ:
                 Hm11_.frequency = Hm11_.last_cmd[3];